# openapi.yml
openapi: "3.0.3"
info:
  title: Iroko API
  description: |
    REST API for Iroko CRIS — a Current Research Information System backed by
    PostgreSQL (source of truth) and Memgraph (graph store). All writes go
    through NodeService which atomically updates both stores.

    Authentication: Bearer JWT tokens obtained via `/v1/auth/token`.
    Most endpoints require authentication. Superuser role required for
    admin and crawler endpoints.
  version: "0.1.0"
  contact:
    name: Iroko CRIS Team
    url: https://sceiba.cu
servers:
  - url: /api
    description: Default base path (configured via BASE_PATH env)
paths:
...
  /v1/cypher/edit/node/properties:
    patch:
      operationId: updateNodeProperties
      summary: Update node properties
      tags: [Cypher Edit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodePropertyUpdate"
      responses:
        "200":
          description: Properties updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditResponse"
        "404":
          description: Node not found
      security:
        - BearerAuth: []
  /v1/cypher/edit/relationships:
    post:
      operationId: createOrUpdateRelationship
      summary: Create or update a relationship
      tags: [Cypher Edit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipUpdate"
      responses:
        "200":
          description: Relationship created/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditResponse"
        "404":
          description: Source or target node not found
      security:
        - BearerAuth: []
    delete:
      operationId: deleteRelationship
      summary: Delete a relationship
      tags: [Cypher Edit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipDeleteRequest"
      responses:
        "200":
          description: Relationship deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditResponse"
        "404":
          description: Source node not found
      security:
        - BearerAuth: []
  /v1/cypher/edit/node/full:
    put:
      operationId: fullNodeEdit
      summary: Full node edit (properties + relationships)
      tags: [Cypher Edit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeEditRequest"
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditResponse"
        "404":
          description: Node not found
      security:
        - BearerAuth: []

  # ──────────────────────────────────────────────
  # Nodes
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/v1/auth/token`
  schemas:
...
    NodePropertyUpdate:
      type: object
      properties:
        iroko_uuid:
          type: string
          format: uuid
        properties:
          type: object
      required: [iroko_uuid, properties]
    RelationshipUpdate:
      type: object
      properties:
        from_uuid:
          type: string
          format: uuid
        to_uuid:
          type: string
          format: uuid
        relation_type:
          type: string
          description: Uppercase relationship type (e.g., HAS_AFFILIATION)
        properties:
          type: object
          nullable: true
      required: [from_uuid, to_uuid, relation_type]
    RelationshipDeleteRequest:
      type: object
      properties:
        from_uuid:
          type: string
          format: uuid
        to_uuid:
          type: string
          format: uuid
        relation_type:
          type: string
      required: [from_uuid, to_uuid, relation_type]
    NodeEditRequest:
      type: object
      properties:
        iroko_uuid:
          type: string
          format: uuid
        properties:
          type: object
        relationships:
          type: array
          items:
            $ref: "#/components/schemas/RelationshipUpdate"
      required: [iroko_uuid, properties, relationships]
    EditResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        updated_properties:
          type: integer
          nullable: true
        updated_relationships:
          type: integer
          nullable: true
        deleted_relationships:
          type: integer
          nullable: true
      required: [success, message]

    # ── Nodes ──
...